Rename : Input 디렉토리 분리 및 import문 수정
