+++ src/Components/Input/index.js
@@ -1,2 +1,1 @@
-export { default as Input } from "./Input";
-export { default as InputWithLabel } from "./InputWithLabel";
+export { default } from "./Input";
