+++ src/Components/index.js
@@ -3,6 +3,7 @@ export { default as Button } from "./Button";
 export { default as Comment } from "./Comment";
 export { default as ImageBox } from "./ImageBox";
 export { default as Input } from "./Input";
+export { default as InputWithLabel } from "./InputWithLabel";
 export { default as NavItem } from "./NavItem";
 export { default as SideBarItem } from "./SideBarItem";
 export { default as SubNavbarItem } from "./SubNavbarItem";
